# Conflicts:
#	src/main/java/org/ggp/base/player/gamer/statemachine/ConfigurableStateMachineGamer.java
#	src/main/java/org/ggp/base/player/gamer/statemachine/MCTS/MctsGamer.java
#	src/main/java/org/ggp/base/player/gamer/statemachine/MCTS/manager/hybrid/GameDependentParameters.java
#	src/main/java/org/ggp/base/player/gamer/statemachine/MCTS/manager/hybrid/HybridMctsManager.java
#	src/main/java/org/ggp/base/player/gamer/statemachine/MCTS/manager/hybrid/SharedReferencesCollector.java
#	src/main/java/org/ggp/base/player/gamer/statemachine/MCTS/manager/hybrid/strategies/aftergame/CompositeAfterGame.java
#	src/main/java/org/ggp/base/player/gamer/statemachine/MCTS/manager/hybrid/strategies/aftergame/SimLimLsiTunerAfterGame.java
#	src/main/java/org/ggp/base/player/gamer/statemachine/MCTS/manager/hybrid/strategies/aftermetagame/SimLimDynamicLsiAfterMetagame.java
#	src/main/java/org/ggp/base/player/gamer/statemachine/MCTS/manager/hybrid/strategies/beforesimualtion/TunerBeforeSimulation.java
#	src/main/java/org/ggp/base/player/gamer/statemachine/MCTS/manager/parameterstuning/UcbMultiPopEvoParametersTuner.java
#	src/main/java/org/ggp/base/player/gamer/statemachine/MCTS/manager/parameterstuning/discretetuners/NaiveParametersTuner.java
#	src/main/java/org/ggp/base/player/gamer/statemachine/MCTS/manager/parameterstuning/discretetuners/SequentialParametersTuner.java
#	src/main/java/org/ggp/base/player/gamer/statemachine/MCTS/manager/parameterstuning/discretetuners/SimLimitedLsiParametersTuner.java
#	src/main/java/org/ggp/base/player/gamer/statemachine/MCTS/manager/parameterstuning/evolution/StandardEvolutionManager.java
#	src/main/java/org/ggp/base/player/gamer/statemachine/MCTS/manager/parameterstuning/evolution/UcbEvolutionManager.java
#	src/main/java/org/ggp/base/player/gamer/statemachine/MCTS/manager/parameterstuning/evolution/mutation/SingleRandomMutation.java
#	src/main/java/org/ggp/base/player/gamer/statemachine/MCTS/manager/parameterstuning/structure/NaivePhaseSettings.java
#	src/main/java/org/ggp/base/player/gamer/statemachine/MCTS/manager/parameterstuning/structure/ParametersManager.java
#	src/main/java/org/ggp/base/player/gamer/statemachine/MCTS/manager/parameterstuning/structure/problemrep/SimLimitedLsiProblemRepresentation.java

+++ src/main/java/csironi/ggp/course/MCTS/MCTNode.java
@@ -197,7 +197,7 @@ public class MCTNode implements OldBackpropagationStrategy {
 		return this.visitedChildren.get(index);
 	}
 
-	public List<Integer> getTerminalGoals() throws GoalDefinitionException, StateMachineException{
+	public List<Double> getTerminalGoals() throws GoalDefinitionException, StateMachineException{
 		return this.stateMachine.getGoals(this.state);
 	}
 
@@ -206,7 +206,7 @@ public class MCTNode implements OldBackpropagationStrategy {
 	}
 
 	@Override
-	public void update(List<Integer> goals){
+	public void update(List<Double> goals){
 
 		int numberOfRoles = this.jointMoves.size();
