+++ src/main/java/csironi/ggp/course/MCTS/selection/OldRandomSelection.java
@@ -46,9 +46,9 @@ public class OldRandomSelection implements OldSelectionStrategy {
 	 * @see csironi.ggp.course.MCTS.selection.SelectionStrategy#select(csironi.ggp.course.MCTS.MCTNode, org.ggp.base.util.statemachine.Role)
 	 */
 	@Override
-	public List<Integer> select(MCTNode node) throws GoalDefinitionException, MoveDefinitionException, TransitionDefinitionException, StateMachineException {
+	public List<Double> select(MCTNode node) throws GoalDefinitionException, MoveDefinitionException, TransitionDefinitionException, StateMachineException {
 
-		List<Integer> goals;
+		List<Double> goals;
 
 		// If the node has no children it means it is a leaf node of the MCTS tree
 		if(node.hasNoChildren()){
