+++ src/main/java/org/ggp/base/player/gamer/statemachine/MCTS/manager/hybrid/strategies/aftergame/TunerAfterGame.java
@@ -37,7 +37,7 @@ public class TunerAfterGame extends AfterGameStrategy {
 	}
 
 	@Override
-	public void afterGameActions(List<Integer> terminalGoals) {
+	public void afterGameActions(List<Double> terminalGoals) {
 		// If the tuner was still tuning we let it set and memorize the best combination of parameters values
 		if(this.parametersTuner.isTuning()){
 			this.parametersTuner.setBestCombinations();
